refactor(feature): add src and types folders
