+++ packages/feature/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from './Feature';
-export * from './Feature';
+export { default } from './types/Feature';
+export * from './types/Feature';
