+++ packages/feature/package.json
@@ -4,6 +4,7 @@
   "author": "Evan Sharp <evan.sharp@gmail.com>",
   "description": "Check if a feature is enabled or disabled before doing something.",
   "main": "index.js",
+  "types": "index.d.ts",
   "bin": {
     "generate-features-json": "bin.js"
   },
@@ -37,7 +38,7 @@
     "react-dom": "^17.0.2"
   },
   "peerDependencies": {
-    "axios": "^0.19.2",
+    "axios": "^0.21.1",
     "react": ">=16.3.0"
   }
 }
